EduEase_Project/backend/requirements.txt
fastapi
uvicorn
google-generativeai
faster-whisper
gtts
yt-dlp
python-dotenv 